backend/requirements.txt
agentscope
fastapi
uvicorn
finnhub-python
openai
httpx
pydantic
python-dotenv